# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/profiles": {
      "get": {
        "operationId": "Profiles_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "ค้นจากชื่อ อีเมล หรือหน่วยงาน",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "USER",
                "TECHNICIAN",
                "ADMIN"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProfileDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ผู้ใช้ที่เคยเข้าระบบนี้",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/{id}/avatar": {
      "get": {
        "operationId": "Profiles_avatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ไฟล์รูป",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "image/jpeg": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/png": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/webp": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "image/jpeg": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/png": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/webp": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "image/jpeg": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/png": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              },
              "image/webp": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "รูปโปรไฟล์ (image/jpeg · image/png · image/webp)",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/{id}": {
      "patch": {
        "operationId": "Profiles_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProfileDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — ชนกฎธุรกิจ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "แต่งตั้ง/ถอดถอนช่างซ่อมบำรุง",
        "tags": [
          "profiles"
        ]
      }
    },
...
  },
  "info": {
    "title": "csmju-repair API",
    "description": "ระบบแจ้งซ่อม — ระบบย่อยของ CSMJU2030 · ทุก endpoint ใต้ /api/v1 ต้องมี Core Hub access token (Authorization: Bearer หรือคุกกี้ core_hub_access_token) · response ห่อด้วย envelope มาตรฐาน",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT (RS256 จาก Core Hub)",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "VALIDATION_ERROR",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "CONFLICT",
          "INTERNAL_ERROR"
        ]
      },
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "ข้อความภาษาไทยที่แสดงให้ผู้ใช้ได้ทันที"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "ProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "id ของโปรไฟล์ในระบบนี้ (ใช้กับ PATCH /profiles/:id)"
          },
          "coreUserId": {
            "type": "string",
            "example": "user-003",
            "description": "claim `sub` จาก Core Hub"
          },
          "email": {
            "type": "string",
            "example": "staff@core.local"
          },
          "coreRole": {
            "type": "string",
            "enum": [
              "student",
              "alumni",
              "staff",
              "admin"
            ]
          },
          "subsystemRole": {
            "type": "string",
            "enum": [
              "USER",
              "TECHNICIAN",
              "ADMIN"
            ],
            "nullable": true,
            "description": "null = core role นี้เข้าระบบไม่ได้แล้ว"
          },
          "isTechnician": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string",
            "example": "สมชาย ใจดี"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "0812345678"
          },
          "workUnit": {
            "type": "string",
            "nullable": true,
            "example": "งานอาคารสถานที่"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "description": "รูปโปรไฟล์ (GET ได้เมื่อเข้าสู่ระบบแล้ว) · null = ยังไม่มีรูป ให้แสดงอักษรย่อแทน"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "coreUserId",
          "email",
          "coreRole",
          "subsystemRole",
          "isTechnician",
          "displayName",
          "phone",
          "workUnit",
          "avatarUrl",
          "lastSeenAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 42
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "isTechnician": {
            "type": "boolean",
            "description": "แต่งตั้ง/ถอดถอนช่างซ่อมบำรุง (เฉพาะผู้ใช้ core role staff)"
          }
        },
        "required": [
          "isTechnician"
        ]
      },
...
    }
  }
}